major refactor: move SUPPORTED_LANGUAGES to serve definition, remove unnecessary dependency injections
