+++ packages/shared/src/services/TranscriptionService.ts
@@ -1,11 +1,139 @@
-import type { SupportedLanguage, WhisperingError } from '@repo/shared';
+import type { WhisperingError } from '@repo/shared';
 import type { Effect } from 'effect';
 import { Context } from 'effect';
+
+/** Supported languages pulled from OpenAI Website: https://platform.openai.com/docs/guides/speech-to-text/supported-languages */
+export const SUPPORTED_LANGUAGES = [
+	'auto',
+	'af',
+	'ar',
+	'hy',
+	'az',
+	'be',
+	'bs',
+	'bg',
+	'ca',
+	'zh',
+	'hr',
+	'cs',
+	'da',
+	'nl',
+	'en',
+	'et',
+	'fi',
+	'fr',
+	'gl',
+	'de',
+	'el',
+	'he',
+	'hi',
+	'hu',
+	'is',
+	'id',
+	'it',
+	'ja',
+	'kn',
+	'kk',
+	'ko',
+	'lv',
+	'lt',
+	'mk',
+	'ms',
+	'mr',
+	'mi',
+	'ne',
+	'no',
+	'fa',
+	'pl',
+	'pt',
+	'ro',
+	'ru',
+	'sr',
+	'sk',
+	'sl',
+	'es',
+	'sw',
+	'sv',
+	'tl',
+	'ta',
+	'th',
+	'tr',
+	'uk',
+	'ur',
+	'vi',
+	'cy',
+] as const;
+
+export type SupportedLanguage = (typeof SUPPORTED_LANGUAGES)[number];
+
+export const SUPPORTED_LANGUAGES_TO_LABEL = {
+	auto: 'Auto',
+	af: 'Afrikaans',
+	ar: 'Arabic',
+	hy: 'Armenian',
+	az: 'Azerbaijani',
+	be: 'Belarusian',
+	bs: 'Bosnian',
+	bg: 'Bulgarian',
+	ca: 'Catalan',
+	zh: 'Chinese',
+	hr: 'Croatian',
+	cs: 'Czech',
+	da: 'Danish',
+	nl: 'Dutch',
+	en: 'English',
+	et: 'Estonian',
+	fi: 'Finnish',
+	fr: 'French',
+	gl: 'Galician',
+	de: 'German',
+	el: 'Greek',
+	he: 'Hebrew',
+	hi: 'Hindi',
+	hu: 'Hungarian',
+	is: 'Icelandic',
+	id: 'Indonesian',
+	it: 'Italian',
+	ja: 'Japanese',
+	kn: 'Kannada',
+	kk: 'Kazakh',
+	ko: 'Korean',
+	lv: 'Latvian',
+	lt: 'Lithuanian',
+	mk: 'Macedonian',
+	ms: 'Malay',
+	mr: 'Marathi',
+	mi: 'Maori',
+	ne: 'Nepali',
+	no: 'Norwegian',
+	fa: 'Persian',
+	pl: 'Polish',
+	pt: 'Portuguese',
+	ro: 'Romanian',
+	ru: 'Russian',
+	sr: 'Serbian',
+	sk: 'Slovak',
+	sl: 'Slovenian',
+	es: 'Spanish',
+	sw: 'Swahili',
+	sv: 'Swedish',
+	tl: 'Tagalog',
+	ta: 'Tamil',
+	th: 'Thai',
+	tr: 'Turkish',
+	uk: 'Ukrainian',
+	ur: 'Urdu',
+	vi: 'Vietnamese',
+	cy: 'Welsh',
+} as const satisfies Record<SupportedLanguage, string>;
+
+export const supportedLanguagesOptions = SUPPORTED_LANGUAGES.map(
+	(lang) => ({ label: SUPPORTED_LANGUAGES_TO_LABEL[lang], value: lang }) as const,
+);
 
 export class TranscriptionService extends Context.Tag('TranscriptionService')<
 	TranscriptionService,
 	{
-		readonly supportedLanguages: { label: string; value: string }[];
 		readonly transcribe: (
 			blob: Blob,
 			options: { apiKey: string; outputLanguage: SupportedLanguage },
